fix: complete wui->config rename left incomplete on develop
