+++ backend/request/views.py
@@ -2,8 +2,10 @@ import itertools
 import json
 import logging
 import os
+from mimetypes import guess_type
+from os.path import basename
 from unicodedata import normalize
-from urllib.parse import urlencode
+from urllib.parse import quote, urlencode
 
 from common.mviews import refresh_batched
 from common.serializers import UserSerializer
@@ -12,6 +14,7 @@ from common.views import CsrfExemptSessionAuthentication, StandardResultsSetPagi
 from django.apps import apps
 from django.conf import settings
 from django.contrib.auth import get_user_model
+from django.contrib.auth.decorators import login_required
 from django.contrib.sites.shortcuts import get_current_site
 from django.core.mail import send_mail
 from django.db import transaction
@@ -1065,3 +1068,51 @@ class ApproveViewSet(viewsets.ModelViewSet):
                 recipient_list=_recipient_list_with_sender_copy(recipient_email),
             )
         return HttpResponseRedirect("/danke")
+
+
+@login_required
+def protected_media(request, *args, **kwargs):
+    """Protected view for media files"""
+
+    allow_download = False
+    url_path = kwargs["url_path"]
+
+    if request.user.is_staff:
+        allow_download = True
+    elif request.user.is_pi:
+        # Master/ PI accounts should be able to access attachments
+        allow_download = (
+            request.user.pi
+            == Request.objects.filter(
+                Q(deep_seq_request=url_path) | Q(files__file=url_path), archived=False
+            )[0].user.pi
+        )
+    else:
+        allow_download = Request.objects.filter(
+            Q(deep_seq_request=url_path) | Q(files__file=url_path),
+            user=request.user,
+            archived=False,
+        ).exists()
+
+    if allow_download:
+        response = HttpResponse()
+
+        # Set file type and encoding
+        mimetype, encoding = guess_type(url_path)
+        response["Content-Type"] = mimetype if mimetype else "application/octet-stream"
+        if encoding:
+            response["Content-Encoding"] = encoding
+
+        # Set internal redirect to protected media
+        response["X-Accel-Redirect"] = f"/protected_media/{url_path}"
+
+        # Set file name
+        file_name = basename(url_path)
+        # Needed for file names that include special, non ascii, characters
+        response["Content-Disposition"] = (
+            f"attachment; filename*=utf-8''{quote(file_name)}"
+        )
+
+        return response
+
+    raise Http404
