+++ backend/requirements/3.13/dev.txt
@@ -31,6 +31,8 @@ cffi==2.1.1 ; platform_python_implementation != 'PyPy'
     # via cryptography
 charset-normalizer==3.5.1
     # via requests
+click==8.5.0
+    # via import-linter
 colorama==0.4.6 ; sys_platform == 'win32'
     # via ipython
 contourpy==1.3.3
@@ -142,6 +144,8 @@ fpdf2==2.8.8
     # via
     #   -c backend/requirements/3.13/base.txt
     #   -r backend/requirements/base.in
+grimp==3.17
+    # via import-linter
 gunicorn==26.2.0
     # via
     #   -c backend/requirements/3.13/base.txt
@@ -159,6 +163,8 @@ idna==3.19
     #   anyio
     #   httpx2
     #   requests
+import-linter==2.15
+    # via -r backend/requirements/dev.in
 inflection==0.5.1
     # via
     #   -c backend/requirements/3.13/base.txt
@@ -285,7 +291,9 @@ requests==2.34.2
     #   django-sql-explorer
     #   tiktoken
 rich==15.0.0
-    # via -r backend/requirements/dev.in
+    # via
+    #   -r backend/requirements/dev.in
+    #   import-linter
 rpds-py==2026.6.3
     # via
     #   -c backend/requirements/3.13/base.txt
@@ -335,6 +343,7 @@ typing-extensions==4.16.0
     # via
     #   -c backend/requirements/3.13/base.txt
     #   anyio
+    #   import-linter
     #   openai
     #   pydantic
     #   pydantic-core
